added my own memory allocator for small objects
git-svn-id: https://only.mawhrin.net/repos/stuff/trunk@301 62a0de91-87f2-0310-aeaa-8eaa5733e38e

+++ fbreader/common/textview/Word.cpp
@@ -21,7 +21,7 @@
 
 #include "Word.h"
 
-Word::Word(const std::string &utf8String, size_t start, size_t size, size_t paragraphOffset) : myData(utf8String), myStart(start), mySize(size), myMark(0), myParagraphOffset(paragraphOffset), myWidth(-1) {
+Word::Word(const std::string &utf8String, size_t start, unsigned short size, size_t paragraphOffset) : myData(utf8String), myStart(start), mySize(size), myMark(0), myParagraphOffset(paragraphOffset), myWidth(-1) {
 	myLength = ZLUnicodeUtil::utf8Length(myData.data() + start, size);
 }
